Troubleshooting search bar
Committing this to local branch so I can safely delete some code that I might want to reference later. Some of the features are working correctly but others aren't. The search results are displaying the correct keys for the provided values, but a variety of other features are lacking.

1. the arrow keys do not work appropriately
2. clicking outside of the search bar does not clear the autocomplete options from the screen
3. if the input field is empty, it does not display the full list of terms (except for when it is initially clicked on)
4. I have not done any testing for mobile devices
5. the enter key still defaults to the first result on the autocomplete options, instead of the one currently highlighted by the user
6. the options do not highlight unless hovered over by the mouse - the arrow keys should cycle through them in the same way

Additionally, the results on the list of terms are not being highlighted correctly.

Made some changes to the color scheme and made the answer box fixed in the middle of the screen instead of moving over the corresponding column.

Closes #2