Fixed keyboard responses for search bar
Added functionality for the up/down arrow keys to change the currently highlighted item in the autocomplete menu, and added core functionality of enter button/clicking the search icon to execute with the user's input appropriately.

Added a dummy header - this should be changed to include the title (Praise the message) as well as tabs for various souls-like games, which pull up separate pages for each.

closes #3, closes #10