debugging Netlify deployment issue
I don't  think I can get away with index.html being in a pages subdirectory - it either needs to be in root, or I need to keep everything else in a subdirectory from index.html, and I'm not about to store everything in a directory called pages. Maybe I'll switch to src at some point.